+++ modules/AWSPowerShell/Cmdlets/SSOOIDC/Basic/Register-SSOOIDCClient-Cmdlet.cs
@@ -28,9 +28,8 @@ using Amazon.SSOOIDC.Model;
 namespace Amazon.PowerShell.Cmdlets.SSOOIDC
 {
     /// <summary>
-    /// Registers a client with Amazon Web Services SSO. This allows clients to initiate device
-    /// authorization. The output should be persisted for reuse through many authentication
-    /// requests.
+    /// Registers a client with AWS SSO. This allows clients to initiate device authorization.
+    /// The output should be persisted for reuse through many authentication requests.
     /// </summary>
     [Cmdlet("Register", "SSOOIDCClient", SupportsShouldProcess = true, ConfirmImpact = ConfirmImpact.Medium)]
     [OutputType("Amazon.SSOOIDC.Model.RegisterClientResponse")]